woo-hoo! Now CollisionTermInfo of trajopt is implemented.
This version successfully uses Bullet collision detector from MoveIt to do the collsion checking and find singed distances. There are a lot of changes in different
files but the most changes are done in collision_terms.cpp

Things left to do:
1. This implementation is hard-coded. But it is easy to change this to a version that the user can enter a CollisionTermInfo.
Currently this is done in trajopt_interface.cpp but it should be defined by user, its parametes should be added to setup.yaml as well.

2. It needs cleanup in the code base.

3. The PlanningScene that motion planners have acces to is a ConsPtr but we need to make changes in RobotState to be able to check collision. To this end,
a child of the original PlanningScene is used through diff() function. For some reason, I can not use "checkCollision" function on this planning scene child.
It turned out if I remove all the objects from the child and add them again, "checkCollision" operates correctly